(feat) removed attendancerecord and condensed to students
