+++ src/public/components/requests/students.js
@@ -20,4 +20,14 @@ const getStudentData = async (studentData) => {
   }
 };
 
-export { storeStudentData, getStudentData };
+const getStudentInCertainClasses = async (classes) => {
+  try {
+    const splitClasses = classes.split(',');
+    const reponse = await axios.post('/getStudentWithCertainClasses', { classes: splitClasses});
+    return reponse.status === 201;
+  } catch (err) {
+    console.warn(err);
+  }
+}
+
+export { storeStudentData, getStudentData, getStudentInCertainClasses };
